Move implementation onto PgConnection.
If we use `&mut PgConnection` as our executor, we can create a `&mut
PgConnection` by either borrowing our pooled connection or by passing in
a borrow of a `Transaction`.

This requires splitting the code into two:
 - Something that has this `&mut Connection` and mutably acts on it in
   order to exeucte queries. This is `Connection<'a>` in this code.
 - Something that holds the `PgPool` and implements the `Repository`.
   This can acquire connections from the pool and then mutably borrow
   them in order to execute queries. This is `Database` in this code.

All of the actual logic exists on `Connection`. We'd test this using our
'wrapping tests in transaction' approach.